+++ packages/nimble-components/src/table-column/text/index.ts
@@ -23,41 +23,36 @@ declare global {
 }
 
 /**
- * The base class for a table column for displaying strings.
+ * The table column for displaying strings.
  */
-abstract class TableColumnTextBase extends TableColumn<TableColumnTextColumnConfig> {
-    public cellRecordFieldNames = ['value'] as const;
-
+export class TableColumnText extends mixinGroupableColumnAPI(
+    mixinFractionalWidthColumnAPI(TableColumn<TableColumnTextColumnConfig>)
+) {
     @attr({ attribute: 'field-name' })
     public fieldName?: string;
 
     @attr
     public placeholder?: string;
 
-    public readonly cellViewTag = tableColumnTextCellViewTag;
-
     public constructor() {
-        super();
-        this.sortOperation = TableColumnSortOperation.localeAwareCaseSensitive;
+        super({
+            cellRecordFieldNames: ['value'],
+            cellViewTag: tableColumnTextCellViewTag,
+            groupHeaderViewTag: tableColumnTextGroupHeaderTag
+        });
+        this.columnInternals.sortOperation = TableColumnSortOperation.localeAwareCaseSensitive;
     }
 
     protected fieldNameChanged(): void {
-        this.dataRecordFieldNames = [this.fieldName] as const;
-        this.operandDataRecordFieldName = this.fieldName;
+        this.columnInternals.dataRecordFieldNames = [this.fieldName];
+        this.columnInternals.operandDataRecordFieldName = this.fieldName;
     }
 
     protected placeholderChanged(): void {
-        this.columnConfig = { placeholder: this.placeholder ?? '' };
+        this.columnInternals.columnConfig = {
+            placeholder: this.placeholder ?? ''
+        };
     }
-}
-
-/**
- * The table column for displaying strings.
- */
-export class TableColumnText extends mixinGroupableColumnAPI(
-    mixinFractionalWidthColumnAPI(TableColumnTextBase)
-) {
-    public groupHeaderViewTag = tableColumnTextGroupHeaderTag;
 }
 
 const nimbleTableColumnText = TableColumnText.compose({